fix: circular reference
Add system/assert.js
Move assert_enum_set to enumeration.js
