+++ src/library/extra_lib.js
@@ -3,7 +3,7 @@ import { define_libfunc, alias_libfunc, define_syntax, define_scmfunc,
          assert_number, assert_integer, assert_real, assert_between, assert_string,
          assert_char, assert_symbol, assert_port, assert_pair, assert_list,
          assert_vector, assert_hashtable, assert_mutable_hashtable, assert_record,
-         assert_record_td, assert_record_cd, assert_enum_set, assert_promise,
+         assert_record_td, assert_record_cd, assert_promise,
          assert_function, assert_closure, assert_procedure, assert_date } from "./infra.js"; 
 import { to_write, to_display, to_write_ss, inspect } from "../system/_writer.js"
 import { Pair, List, array_to_list, deep_array_to_list } from "../system/pair.js"
