+++ src/library/srfi.js
@@ -3,7 +3,7 @@ import { define_libfunc, alias_libfunc, define_syntax, define_scmfunc,
          assert_number, assert_integer, assert_real, assert_between, assert_string,
          assert_char, assert_symbol, assert_port, assert_pair, assert_list,
          assert_vector, assert_hashtable, assert_mutable_hashtable, assert_record,
-         assert_record_td, assert_record_cd, assert_enum_set, assert_promise,
+         assert_record_td, assert_record_cd, assert_promise,
          assert_function, assert_closure, assert_procedure, assert_date, assert, deprecate } from "./infra.js"; 
 import { Pair, array_to_list, deep_array_to_list } from "../system/pair.js"
 import { Port } from "../system/port.js"
